feat: asciidoctor-diagram support
Enables `asciidoctor-diagram` support for generating images as defined inline in
the source file. The resulting images are captured afterwards in a new task
named `collect-images`, to stick to the basic content-task structure for the
base asciidoctor task.

Previous code was refactored to have stronger defaults for options and metadata
handling.